rename: auth 디렉토리 생성 및 관련 코드 이동
