+++ src/main/kotlin/org/scaleadvisor/backend/global/auth/dto/LogoutRequest.kt
@@ -1,4 +1,4 @@
-package org.scaleadvisor.backend.user.dto
+package org.scaleadvisor.backend.global.auth.dto
 
 data class LogoutRequest (
     val refreshToken: String
